// backend/src/common/enums/permiso-sistema.enum.ts
export enum PermisoSistema {
  LOGISTICA_VER = 'LOGISTICA_VER',
  LOGISTICA_GESTIONAR = 'LOGISTICA_GESTIONAR',
  DASHBOARD_VER = 'DASHBOARD_VER',
  USUARIOS_VER = 'USUARIOS_VER',
  USUARIOS_GESTIONAR = 'USUARIOS_GESTIONAR',
  SUCURSALES_VER = 'SUCURSALES_VER',
  SUCURSALES_GESTIONAR = 'SUCURSALES_GESTIONAR',
  MEDIOS_PAGO_VER = 'MEDIOS_PAGO_VER',
  MEDIOS_PAGO_GESTIONAR = 'MEDIOS_PAGO_GESTIONAR',
  PERSONAS_VER = 'PERSONAS_VER',
  PERSONAS_GESTIONAR = 'PERSONAS_GESTIONAR',
}